Implement Task 1, parts A and B
This commit adds code fulfilling most of the requirements of Task 1, part A
and Task 1, part B (only a few statistical tests to be done at end of part B
remain).

The code in this task can be reused in later ones, so committing at this point.

Note that this commit also includes an implementation of the AES-128 CBC
algorithm, but only the encryption part, as that was needed to generate random
numbers using the AES algorithm. The code includes a unit test to ensure things
work and a script to be used with a profiler. This module should not require
further change.

Author: Prajjwal Mishra